+++ test/src/cognitect/aws/ec2_metadata_utils_test.clj
@@ -2,7 +2,7 @@
 ;; All rights reserved.
 
 (ns cognitect.aws.ec2-metadata-utils-test
-  (:require [clojure.test :refer [deftest is use-fixtures]]
+  (:require [clojure.test :refer [deftest is testing use-fixtures]]
             [clojure.core.async :as a]
             [cognitect.aws.http :as http]
             [cognitect.aws.client.shared :as shared]
@@ -34,3 +34,8 @@
                              (doto (a/promise-chan)
                                (a/>!! {:cognitect.anomalies/category :cognitect.anomalies/busy})))]
     (is (nil? (ec2-metadata-utils/get-ec2-instance-region *http-client*)))))
+
+(deftest request-map
+  (testing "port"
+    (is (= 443 (:server-port (#'ec2-metadata-utils/request-map (java.net.URI/create "https://169.254.169.254")))))
+    (is (= 80 (:server-port (#'ec2-metadata-utils/request-map (java.net.URI/create "http://169.254.169.254")))))))