+++ CHANGES.md
@@ -1,5 +1,8 @@
 # aws-api
 
+## DEV
+* fix URI parsing bug in ec2-metadata-utils (h/t @stevebuik)
+
 ## 0.8.630 / 2022-12-05
 * log the credentials provider that succeeded [#217](https://github.com/cognitect-labs/aws-api/issues/217)
 * add keyword access to :api on the test double client [#224](https://github.com/cognitect-labs/aws-api/issues/224)